streamlit/requirements.txt
streamlit==1.18.0
vertexai==1.71.1
google-api-python-client==2.84.0
tenacity>=8.2.3,<9.0.0
langchain==0.3.9
langchain-community==0.3.9
chromadb
openai==0.27.7
google-auth==2.21.0
python-dotenv==1.0.0
pydantic==2.9.2
numpy==1.24.2
altair==4.2.2
tiktoken
